[GH-37] Implement slash command autocomplete
This commit adds autocomplete prompts for all of the subcommands offered by the plugin.
Adds a new `join` subcommand which is an alias for `/webex <@username>/<room id>. This was required as the Autocomplete library does not allow for a command to have both subcommands and arguments.

Fixes https://github.com/mattermost/mattermost-plugin-webex/issues/37